requirements.txt
adblockparser==0.7
amqp==2.2.2
asgi-redis==1.0.0
asgiref==2.2.0
async-timeout==2.0.1
attrdict==2.0.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
autopep8==1.3.4
awscli==1.14.63
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==2.1.3
boto==2.48.0
boto3==1.6.11
botocore==1.9.16
bs4==0.0.1
cachetools==2.0.1
celery==4.1.0
certifi==2018.1.18
channels==2.0.2
chardet==3.0.4
colorama==0.3.7
constantly==15.1.0
cytoolz==0.9.0.1
daphne==2.1.0
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11
django-admin-honeypot==1.0.0
django-channels==0.7.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-easy-audit==1.0
django-embed-video==1.1.2
django-ipware==2.0.1
django-progressive-web-app==0.1.1
django-rest-framework==0.1.0
django-storages==1.6.5
djangorestframework==3.7.7
docker-pycreds==0.2.2
docutils==0.14
ecdsa==0.13
entrypoints==0.2.3
esptool==2.3.1
eth-account==0.2.3
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
extraction==0.2
google-api-python-client==1.6.5
google-auth==1.4.1
googlemaps==2.5.1
gunicorn==19.7.1
hexbytes==0.1.0
html5lib==1.0.1
httplib2==0.10.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kombu==4.1.0
lxml==4.2.1
MarkupSafe==1.0
mistune==0.8.3
msgpack-python==0.5.6
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.2
oauth2client==4.1.2
oauthlib==2.0.6
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
pyaes==1.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyCmdMessenger==0.2.4
pycodestyle==2.3.1
pycryptodome==3.6.1
Pygments==2.2.0
PyJWT==1.6.1
pyserial==3.4
python-dateutil==2.6.1
python-http-client==3.0.0
python3-openid==3.1.0
pytz==2018.3
pywinpty==0.5.4
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rlp==1.0.1
rsa==3.4.2
rules==1.3
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
selenium==3.11.0
Send2Trash==1.5.0
sendgrid==5.3.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
summary-extraction==0.2
tabula-py==1.0.1
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
Twisted==17.9.0
txaio==2.9.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
vpnotebook==0.1.3
vpython==7.4.4
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.47.0
whitenoise==3.3.1
widgetsnbextension==3.2.1
xlrd==1.1.0
zope.interface==4.4.3
python-docx